Attempt to support tags in ToLogstash
There was some code to support tags in ToLogstash, but it seemed incomplete, in
that there was no moving tags from @fields after moving everything to @fields.

Move them back out of fields, so the adding tags and default tags work as
intended. Add tests for passing bot a string in tags, and an array ref.

We handle the moving outside of the %MAP, because the loop stringifies, and
making it not stringify complicated it more than the extra if clause (in my
opinion).